requirements.txt
Django
django-extensions
pandas
djangorestframework
drf-yasg
coreapi
pyyaml